% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_ghost_matrix.R
\name{BlockGhostMatrix}
\alias{BlockGhostMatrix}
\title{Creates an instance of a BlockGhostMatrix.}
\usage{
BlockGhostMatrix(
  blocks = list(),
  matrices = list(),
  vectors = list(),
  n.groups = c()
)
}
\arguments{
\item{blocks}{list of matrices of the same type.
If specified, matrices, vectors, n.groups will be ignored.
Otherwise, the three will be used instead.}

\item{matrices}{list of dense matrices to construct each GhostMatrix.}

\item{vectors}{list of dense vectors to construct each GhostMatrix.}

\item{n.groups}{vector of the number of groups for each GhostMatrix.
If only one value is specified, it is assumed to be
the same value for each GhostMatrix.}
}
\description{
This matrix represents a block diagonal matrix where each
block matrix is a GhostMatrix.
}
