% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_group_ghost_matrix.R
\name{BlockGroupGhostMatrix}
\alias{BlockGroupGhostMatrix}
\title{Creates an instance of a BlockGroupGhostMatrix.}
\usage{
BlockGroupGhostMatrix(
  blocks = list(),
  off.diagonals = list(),
  diagonals = list(),
  n.groups = c()
)
}
\arguments{
\item{blocks}{list of matrices of the same type.
If specified, the other arguments will be ignored.
Otherwise, the other arguments will be used.}

\item{off.diagonals}{list of dense off-diagonal blocks to construct each GroupGhostMatrix.}

\item{diagonals}{list of dense non-off-diagonal blocks to construct each GroupGhostMatrix.}

\item{n.groups}{vector of the number of groups for each GroupGhostMatrix.
If only one value is specified, it is assumed to be
the same value for each GroupGhostMatrix.}
}
\description{
This matrix represents a block diagonal matrix where each
block matrix is a GroupGhostMatrix.
}
