% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{BlockMatrix}
\alias{BlockMatrix}
\title{Creates an instance of a BlockMatrix.}
\usage{
BlockMatrix(blocks)
}
\arguments{
\item{blocks}{list of matrices of the same type.}
}
\description{
Creates an instance of a BlockMatrix.
}
