% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_block_group_ghost_matrix.R
\name{dim.Rcpp_BlockBlockGroupGhostMatrix__}
\alias{dim.Rcpp_BlockBlockGroupGhostMatrix__}
\title{Overload for BlockBlockGroupGhostMatrix dimension.}
\usage{
\method{dim}{Rcpp_BlockBlockGroupGhostMatrix__}(x)
}
\arguments{
\item{x}{BlockBlockGroupGhostMatrix object.}
}
\description{
Overload for BlockBlockGroupGhostMatrix dimension.
}
