% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_ghost_matrix.R
\name{dim.Rcpp_BlockGhostMatrix__}
\alias{dim.Rcpp_BlockGhostMatrix__}
\title{Overload for BlockGhostMatrix dimension.}
\usage{
\method{dim}{Rcpp_BlockGhostMatrix__}(x)
}
\arguments{
\item{x}{BlockGhostMatrix object.}
}
\description{
Overload for BlockGhostMatrix dimension.
}
