% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_group_ghost_matrix.R
\name{dim.Rcpp_BlockGroupGhostMatrix__}
\alias{dim.Rcpp_BlockGroupGhostMatrix__}
\title{Overload for BlockGroupGhostMatrix dimension.}
\usage{
\method{dim}{Rcpp_BlockGroupGhostMatrix__}(x)
}
\arguments{
\item{x}{BlockGroupGhostMatrix object.}
}
\description{
Overload for BlockGroupGhostMatrix dimension.
}
