% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost_matrix.R
\name{dim.Rcpp_GhostMatrix__}
\alias{dim.Rcpp_GhostMatrix__}
\title{Overload for GhostMatrix dimension.}
\usage{
\method{dim}{Rcpp_GhostMatrix__}(x)
}
\arguments{
\item{x}{GhostMatrix object.}
}
\description{
Overload for GhostMatrix dimension.
}
