% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ghost_matrix.R
\name{dim.Rcpp_GroupGhostMatrix__}
\alias{dim.Rcpp_GroupGhostMatrix__}
\title{Overload for GroupGhostMatrix dimension.}
\usage{
\method{dim}{Rcpp_GroupGhostMatrix__}(x)
}
\arguments{
\item{x}{GroupGhostMatrix object.}
}
\description{
Overload for GroupGhostMatrix dimension.
}
