% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostbasil.R
\name{fit_basil}
\alias{fit_basil}
\title{Fits PGR objective with BASIL framework.}
\usage{
fit_basil(
  A,
  y,
  s,
  n_knockoffs = 0,
  n_lambdas = 100,
  n_lambdas_iter = 10,
  strong_size = 1000,
  delta_strong_size = 100,
  n_iters = 100,
  max_cds = 1000,
  thr = 1e-14
)
}
\arguments{
\item{A}{data covariance matrix (currently must be a dense matrix).}

\item{y}{response vector.}

\item{s}{regularization to shrink A towards identity ((1-s) * A + s * I).}

\item{n_knockoffs}{number of knockoffs used to construct A and y (default: 0).}

\item{n_lambdas}{number of lambdas to compute solutions for. Currently there is no early-stopping rule.}

\item{n_lambdas_iter}{maximum number of lambdas to compute strong set solutions for at each iteration of the BASIL algorithm.}

\item{strong_size}{maximum number of strong set variables to append at each iteration of the BASIL algorithm.}

\item{delta_strong_size}{number of strong set variables to increase strong_size at KKT failure of first lambda.}

\item{n_iters}{maximum number of iterations of BASIL.}

\item{max_cds}{maximum number of total coordinate descents.}

\item{thr}{convergence threshold.}
}
\description{
Fits PGR objective with BASIL framework.
}
