% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost_matrix.R
\name{GhostMatrix}
\alias{GhostMatrix}
\title{Creates an instance of a GhostMatrix.}
\usage{
GhostMatrix(Sigma, D, n.groups)
}
\arguments{
\item{Sigma}{covariance matrix of original features.}

\item{D}{diagonal to subtract from Sigma in the off-diagonal blocks.}

\item{n.groups}{number of groups (e.g. 1 knockoff means 2 groups).}
}
\description{
Creates an instance of a GhostMatrix.
}
