% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Computes the objective that is minimized by ghostbasil
\deqn{
 \frac{1-s}{2} \beta^\top A \beta - \beta^\top r + \frac{s}{2} ||\beta||_2 + \lambda ||\beta||_1
}}
\usage{
objective(A, r, s, lmda, beta)
}
\arguments{
\item{A}{covariance matrix.}

\item{r}{covariance between covariates and response.}

\item{s}{regularization strength of A towards identity.}

\item{lmda}{lambda regularization value}

\item{beta}{vector of values.}
}
\description{
Computes the objective that is minimized by ghostbasil
\deqn{
 \frac{1-s}{2} \beta^\top A \beta - \beta^\top r + \frac{s}{2} ||\beta||_2 + \lambda ||\beta||_1
}
}
