% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_group_ghost_matrix.R
\name{BlockGroupGhostMatrix}
\alias{BlockGroupGhostMatrix}
\title{Creates an instance of a BlockGroupGhostMatrix.}
\usage{
BlockGroupGhostMatrix(Sigma, D, n.groups)
}
\arguments{
\item{Sigma}{covariance matrix of original features (dense matrix) minus D.}

\item{D}{matrix to add to Sigma on the diagonal blocks (BlockMatrix).}

\item{n.groups}{number of groups (e.g. 1 knockoff means 2 groups).}
}
\description{
This matrix has the block-form of Sigma on the off-diagonal
and Sigma + D on the diagonal with D begin block-diagonal.
For example, if n.groups is 2, then GhostMatrix represents a matrix of the form
\deqn{
 \begin{bmatrix}
     \Sigma+D & \Sigma \\
     \Sigma & \Sigma+D
 \end{bmatrix}
}
}
