% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{BlockMatrix}
\alias{BlockMatrix}
\title{Creates an instance of a BlockMatrix.}
\usage{
BlockMatrix(blocks)
}
\arguments{
\item{blocks}{list of dense matrices.}
}
\description{
This matrix represents a block diagonal matrix where each
block matrix is a dense matrix.
}
