% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostbasil.R
\name{ghostbasil}
\alias{ghostbasil}
\title{Fits PGR objective with BASIL framework.}
\usage{
ghostbasil(
  A,
  r,
  s,
  user.lambdas = c(),
  max.lambdas = 100,
  lambdas.iter = 10,
  strong.size = 1000,
  delta.strong.size = 100,
  max.strong.size = 10000,
  max.cds = 1e+05,
  thr = 1e-07,
  n.threads = -1
)
}
\arguments{
\item{A}{data covariance matrix (currently must be a dense matrix).}

\item{r}{correlation vector.}

\item{s}{regularization to shrink A towards identity ((1-s) * A + s * I).}

\item{user.lambdas}{user-specified sequence of lambdas. Will be sorted in decreasing order if not sorted already.}

\item{max.lambdas}{maximum number of lambdas to compute solutions for.}

\item{lambdas.iter}{number of lambdas to compute strong set solutions for at each iteration of the BASIL algorithm.
Internally, capped at max.lambdas.}

\item{strong.size}{initial number of strong set variables to include.
Internally, capped at max.strong.size.}

\item{delta.strong.size}{number of strong set variables to include at every iteration of the BASIL algorithm.
Internally, capped at number of non-strong variables at every iteration of BASIL.}

\item{max.strong.size}{maximum number of strong set size.
Internally, capped at number of features.}

\item{max.cds}{maximum number of coordinate descents per BASIL iteration.}

\item{thr}{coordinate descent convergence threshold.}
}
\description{
Fits PGR objective with BASIL framework.
}
