% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Computes the PGR objective that is minimized by ghostbasil.}
\usage{
objective(A, r, s, lmda, beta)
}
\arguments{
\item{A}{covariance matrix. Currently, only supports dense matrix.}

\item{r}{correlation vector.}

\item{s}{regularization strength of A towards identity.}

\item{lmda}{lambda regularization value.}

\item{beta}{vector of coefficients}
}
\description{
The objective is given by:
\deqn{
 \frac{1-s}{2} \beta^\top A \beta - \beta^\top r + \frac{s}{2} ||\beta||_2 + \lambda ||\beta||_1
}
}
