% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Computes the PGR objective that is minimized by ghostbasil.}
\usage{
objective(A, r, penalty, alpha, lmda, beta)
}
\arguments{
\item{A}{covariance matrix. Currently, only supports dense and sparse matrices.}

\item{r}{correlation vector.}

\item{penalty}{penalty vector p.}

\item{alpha}{elastic net proportion.}

\item{lmda}{lambda regularization value.}

\item{beta}{vector of coefficients}
}
\description{
The objective is given by:
\deqn{
 \frac{1}{2} \beta^\top A \beta - \beta^\top r 
     + \lambda \sum\limits_i p_i \left(\alpha |\beta_i| + \frac{1-\alpha}{2} \beta_i^2 \right)
}
}
